// web/platform-admin-web/src/api/generated/model/getPendingTermsAgreementResponseDto.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * AI Base App Platform Service API
 * The AI Base App Platform Service API documentation
 * OpenAPI spec version: 1.0
 */
import type { GetPendingTermsDocumentResponseDto } from './getPendingTermsDocumentResponseDto';
import type { GetPendingTermsVersionResponseDto } from './getPendingTermsVersionResponseDto';

export interface GetPendingTermsAgreementResponseDto {
  /** 약관 문서 정보 */
  document: GetPendingTermsDocumentResponseDto;
  /** 현재 효력 중인 최신 버전 정보 */
  currentVersion: GetPendingTermsVersionResponseDto;
}